+++ src/Rojan.Desktop.Infrastructure/DependencyInjection/ServiceCollectionExtensions.cs
@@ -1,5 +1,7 @@
 using Microsoft.Extensions.DependencyInjection;
+using Rojan.Desktop.Domain.Customers;
 using Rojan.Desktop.Domain.Dashboard;
+using Rojan.Desktop.Infrastructure.Customers;
 using Rojan.Desktop.Infrastructure.Dashboard;
 
 namespace Rojan.Desktop.Infrastructure.DependencyInjection;
@@ -13,6 +15,7 @@ public static class ServiceCollectionExtensions
     public static IServiceCollection AddInfrastructure(this IServiceCollection services)
     {
         services.AddSingleton<IDashboardRepository, FakeDashboardRepository>();
+        services.AddSingleton<ICustomerRepository, FakeCustomerRepository>();
         return services;
     }
 }